+++ R/rgl.R
@@ -5,6 +5,9 @@
 #' @inheritParams save_piece_obj
 #' @param lit logical, specifying if rgl lighting calculation should take place
 #' @param shininess Properties for rgl lighting calculation
+#' @param textype Use \code{"rgba"} when png texture (may) also have alpha transparency.
+#'                Use \code{"rgb"} when sure texture will not have alpha transparency
+#'                (in particular \code{rgl}'s WebGL export will likely work better).
 #' @return A numeric vector of rgl object IDs.
 #' @examples
 #' if ((Sys.getenv("TRAVIS") == "") && require("rgl")) {
@@ -17,13 +20,15 @@
 #' }
 #' @export
 #' @seealso See \code{\link[rgl]{rgl}} for more information about the \code{rgl} package.
+#'          See \code{\link[rgl]{rgl.material}} for more info about setting \code{rgl} material properties.
 #'          See \code{\link{geometry_utils}} for a discussion of the 3D rotation parameterization.
 piece3d <- function(piece_side = "tile_back", suit = NA, rank = NA, cfg = pp_cfg(), # nolint
                            x = 0, y = 0, z = NA,
                            angle = 0, axis_x = 0, axis_y = 0,
                            width = NA, height = NA, depth = NA,
                            envir = NULL, ..., scale = 1, res = 72,
-                           alpha = 1.0, lit = FALSE, shininess = 50.0) {
+                           alpha = 1.0, lit = FALSE,
+                           shininess = 50.0, textype = "rgba") {
     if (!requireNamespace("rgl", quietly = TRUE)) {
         stop("You need to install the suggested package rgl to use 'piece3d'.",
              "Use 'install.packages(\"rgl\")'")
@@ -46,6 +51,7 @@ piece3d <- function(piece_side = "tile_back", suit = NA, rank = NA, cfg = pp_cfg
     alpha <- rep(alpha, length.out = nn)
     lit <- rep(lit, length.out = nn)
     shininess <- rep(shininess, length.out = nn)
+    textype <- rep(textype, length.out = nn)
 
     cfg <- get_cfg(cfg, envir)
     cfg <- rep(c(cfg), length.out = nn)
@@ -55,7 +61,8 @@ piece3d <- function(piece_side = "tile_back", suit = NA, rank = NA, cfg = pp_cfg
                      angle[i], axis_x[i], axis_y[i],
                      width[i], height[i], depth[i],
                      scale = scale[i], res = res,
-                     alpha = alpha[i], lit = lit[i], shininess = shininess[i])
+                     alpha = alpha[i], lit = lit[i],
+                     shininess = shininess[i], textype = textype[i])
     })
     do.call(c, l)
 }
@@ -65,15 +72,17 @@ rgl_piece_helper <- function(piece_side = "tile_back", suit = NA, rank = NA, cfg
                            angle = 0, axis_x = 0, axis_y = 0,
                            width = NA, height = NA, depth = NA,
                            scale = 1, res = 72,
-                           alpha = 1, lit = FALSE, shininess = 50.0) {
+                           alpha = 1, lit = FALSE,
+                           shininess = 50.0, textype = "rgba") {
     if (scale == 0 || alpha == 0) return(invisible(numeric(0)))
     obj <- save_piece_obj(piece_side, suit, rank, cfg,
                         x = x, y = y, z = z,
                         angle = angle, axis_x = axis_x, axis_y = axis_y,
                         width = width, height = height, depth = depth,
                         scale = scale, res = res)
-    material <- list(color = "white", alpha = alpha, lit = lit, shininess = shininess,
-                     texture = obj$png, textype = "rgba")
+    material <- list(color = "white", alpha = alpha,
+                     lit = lit, shininess = shininess,
+                     texture = obj$png, textype = textype)
     mesh <- suppressWarnings(rgl::readOBJ(obj$obj, material = material))
     invisible(as.numeric(rgl::shade3d(mesh)))
 }